+++ src/org/freeforums/geforce/geffy/main/Reference.java
@@ -4,6 +4,9 @@ import org.freeforums.geforce.geffy.commands.CommandGiveCookie;
 import org.freeforums.geforce.geffy.commands.CommandHelp;
 import org.freeforums.geforce.geffy.commands.CommandRate;
 import org.freeforums.geforce.geffy.commands.ICommand;
+import org.freeforums.geforce.geffy.misc.TS3ActionListener;
+
+import com.github.theholywaffle.teamspeak3.TS3Query;
 
 public class Reference {
 	
@@ -11,10 +14,17 @@ public class Reference {
 	public static final String commandPrefix = "~";
 	
 	public static ICommand[] commands = new ICommand[]{new CommandHelp(), new CommandRate(), new CommandGiveCookie()};
-    public static String[] joinedChannels = new String[]{"#GeforceMods", "Geforce", "#bl4ckscor3", "#Razz"};	
-    public static String[] ignoredUsers = new String[]{"GeffyBot", "Geffy"};	
-    
+    public static String[] joinedChannels = new String[]{"#GeforceMods", "#Geforce", "#BreakInBad", "#bl4ckscor3", "#Razz"};	
+    public static String[] ignoredUsers = new String[]{"Geffy"};
+        
     //Other things.
 	public static String[] cookieTypes = new String[]{"chocolate chip", "peanut butter", "raisin", "oatmeal", "white fudge", "sugar", "frosted", "Oreo"};
-	public static String scVersion = "v1.7.4.1"; //TODO: Change this with every SecurityCraft update
+	public static String scVersion = ""; 
+	public static String scBetaVersion = ""; 
+	public static String scBetaVersionOf = ""; 
+	public static String scBetaDownloadLink = ""; 
+
+	public static TS3Query ts3Query;
+	public static TS3ActionListener ts3EventListener;
+
 }