+++ src/org/freeforums/geforce/geffy/misc/MessageHandler.java
@@ -1,26 +1,31 @@
 package org.freeforums.geforce.geffy.misc;
 
-import java.io.File;
-import java.util.Random;
+import java.util.HashMap;
+import java.util.Iterator;
 
 import org.freeforums.geforce.geffy.main.Bot;
-import org.freeforums.geforce.geffy.utils.FileUtils;
 import org.freeforums.geforce.geffy.utils.Utils;
 import org.pircbotx.hooks.events.MessageEvent;
 
 public class MessageHandler {
 	
-	private static final String[] messagesToRespondTo = new String[]{"new commit to master"};
-
+	public static HashMap<String, String[]> messagesToRespondTo = new HashMap<String, String[]>();
+	public static String[] newCommitToMessages = new String[]{"Oh look, a new commit!", "G17HU8-C0MM17 missile incoming, starboard side! Direct hit!",
+			"G17HU8-C0MM17 missile incoming, port side! Direct hit!", "G17HU8-C0MM17 missile incoming, starboard side! Launching R3V3R7 flare!",
+			"G17HU8-C0MM17 missile incoming, port side! Launching R3V3R7 flare!"};
+	
 	public static void tryRespondingToMessage(MessageEvent<Bot> event) {
-		for(String s : messagesToRespondTo){
-			if(event.getMessage().toLowerCase().contains(s)){
-				File file = new File("messages/" + s + ".txt");
-				int lines = FileUtils.getNumberOfLinesInFile("messages/" + s + ".txt");
-				Utils.sendMessage(event, FileUtils.readLineFromFile("messages/" + s + ".txt", new Random().nextInt(lines + 1)));
-				break;
-			}
+		Iterator<String> keys = messagesToRespondTo.keySet().iterator();
+		
+		while(keys.hasNext()){
+			String key = keys.next();
+			if(event.getMessage().toLowerCase().contains(key)){
+				String[] messages = messagesToRespondTo.get(key);
+	
+				String message = messages[Utils.getRandomNumber(messages.length, true)];
+				Utils.sendMessage(event, message);
+			}	
 		}
-	}
+	}	
 
 }